移动原Java项目目录
