+++ VirusBroadcast_by_Java/src/Bed.java
