+++ VirusBroadcast_by_Java/src/Constants.java
