+++ VirusBroadcast_by_Java/src/Hospital.java
