+++ VirusBroadcast_by_Java/src/MoveTarget.java
