+++ VirusBroadcast_by_Java/src/Person.java
