+++ VirusBroadcast_by_Java/src/PersonPool.java
