+++ VirusBroadcast_by_Java/src/Virus.java
